THIS COMMIT BREAKS PROTOMINE, WORK IN PROGRESS; get r103 for previous stable version.

+++ Makefile
@@ -14,8 +14,10 @@
 ## permissions and limitations under the License.
 ##
 
-
 #### THIS IS IN THE PROTOMINE WORKING DIRECTORY
+
+# if set to true, will treat protomine config as a scratch file
+STILL_IN_TESTING=true
 
 UI=database/ui
 DOC=database/doc
@@ -59,7 +61,7 @@ clobber: clean
 	rm -f database/objects/*
 	rm -f database/relations/*
 	rm -f database/tags/* # leave logs alone
-	rm -f protomine-config.pl
+	$(STILL_IN_TESTING) && rm -f protomine-config.pl
 	rm -f minecode.ps
 
 ###
@@ -115,7 +117,7 @@ errs:
 ###
 
 protomine-config.pl: generate-config.sh
-	./generate-config.sh > protomine-config.pl
+	$(STILL_IN_TESTING) && ./generate-config.sh > protomine-config.pl
 	chmod 755 protomine-config.pl
 
 ###
